fix: Improve Z80 code generation efficiency and add assembly peephole optimization
- Fixed inefficient LD D,H; LD E,L; EX DE,HL pattern in subtraction operations
- Added assembly-level peephole optimization pass to Z80 backend
- Implemented canonical register transfer optimizations:
  - LD L,E; LD H,D => EX DE,HL (more efficient)
  - LD H,D; LD L,E => EX DE,HL (more efficient)
  - LD D,H; LD E,L; EX DE,HL => (eliminated - no-op)
  - Double EX DE,HL => (eliminated - cancels out)
- Fixed regex patterns in assembly peephole optimizer (Go doesn't support backreferences)
- Integrated peephole optimization into Z80 backend when -O flag is used

+++ minzc/pkg/tas/tas_compression.go
@@ -155,7 +155,7 @@ func (c *TASCompression) ReconstructState(targetFrame int64) (*StateSnapshot, er
 	
 	// Apply deltas up to target frame
 	for _, delta := range c.deltaFrames {
-		if delta.Cycle > state.Cycle && int64(delta.Cycle) <= uint64(targetFrame) {
+		if delta.Cycle > state.Cycle && delta.Cycle <= uint64(targetFrame) {
 			c.applyDelta(&state, delta)
 		}
 	}
